requirements.txt
Django==4.0.6
django-ninja==0.17.0
psycopg2-binary
gunicorn
email-validator==1.2.1
pydantic==1.9.1
djangorestframework-simplejwt==5.2.0
django-environ